+++ xenon-common/src/main/java/com/vmware/xenon/common/FactoryService.java
@@ -38,6 +38,17 @@ import com.vmware.xenon.services.common.ServiceUriPaths;
  */
 public abstract class FactoryService extends StatelessService {
 
+    public static Service create(Class<? extends Service> childServiceType,
+            Class<? extends ServiceDocument> childServiceDocumentType) {
+        FactoryService fs = new FactoryService(childServiceDocumentType) {
+            @Override
+            public Service createServiceInstance() throws Throwable {
+                return childServiceType.newInstance();
+            }
+        };
+        return fs;
+    }
+
     public static final int SELF_QUERY_RESULT_LIMIT = 1000;
     private static final long SELF_QUERY_TIMEOUT_MINUTES = 60;
     private EnumSet<ServiceOption> childOptions;